pyproject.toml
[tool.poetry]
name = "server-stats"
version = "0.1.0"
description = ""
authors = []

[tool.poetry.dependencies]
python = ">=3.8.10,<=3.10"
fastapi = "^0.79.0"
psutil = "^5.9.1"
uvicorn = {extras = ["standard"], version = "^0.18.2"}
Jinja2 = "^3.1.2"
websockets = "^10.3"
python-dotenv = "^0.20.0"

[tool.poetry.dev-dependencies]
black = "^22.6.0"
pytest = "^7.1.2"
pytest-asyncio = "^0.19.0"
httpx = "^0.23.0"
requests = "^2.28.1"
trio = "^0.21.0"
asyncio = "^3.4.3"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
filterwarnings = [
    "ignore::DeprecationWarning"
]
